+++ tests/bounds/IfElse/pattern_f/IfElse.c
@@ -1,8 +1,8 @@
 #include "IfElse.h"
 int64_t f(int64_t x){
 	int64_t r = 0;
-	uint16_t _2 = 0;
-	uint16_t _3 = 0;
+	int64_t _2 = 0;
+	int64_t _3 = 0;
 	int64_t _4 = 0;
 	int64_t _5 = 0;
 	int64_t _6 = 0;
@@ -39,27 +39,27 @@ blklab2:;
 
 int main(int argc, char** args){
 	int64_t a = 0;
-	uint64_t b = 0;
+	int64_t b = 0;
 	int64_t c = 0;
 	int64_t d = 0;
 	int64_t _5 = 0;
-	uint16_t _6 = 0;
+	int64_t _6 = 0;
 	int64_t _7 = 0;
 	void* _8;
-	uint64_t _10 = 0;
-	uint16_t _11 = 0;
+	int64_t _10 = 0;
+	int64_t _11 = 0;
 	int64_t _12 = 0;
 	void* _13;
 	int64_t _15 = 0;
-	uint16_t _16 = 0;
+	int64_t _16 = 0;
 	int64_t _17 = 0;
 	void* _18;
 	int64_t _20 = 0;
-	uint16_t _21 = 0;
+	int64_t _21 = 0;
 	void* _22;
 	int64_t _24 = 0;
-	uint16_t _25 = 0;
-	int16_t _26 = 0;
+	int64_t _25 = 0;
+	int64_t _26 = 0;
 	void* _27;
 	//const %6 = 1 : int
 	_6 = 1;
